requirements.txt
audioread==3.0.1
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.4
imbalanced-learn==0.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.3
msgpack==1.0.7
numba==0.58.1
numpy==1.26.1
packaging==23.2
platformdirs==3.11.0
pooch==1.8.0
pycparser==2.21
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
spotipy==2.23.0
threadpoolctl==3.2.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
